Move byte stream controller and BYOB request to separate file
